user routes all work fine now on backend
