+++ controllers/paws_controller.js
@@ -53,17 +53,15 @@ router.post("/users/create", function(req, res) {
     last_name:req.body.last_name,
     user_name: req.body.user_name,
     password:req.body.password,
-    phone: req.body.phone,
-    email:req.body.email,
-    
-    
+    email:req.body.email, 
     // lat: req.body.lat,
     // long: req.body.long
   }).then(function(dbUser) {
       console.log(dbUser);
       // res.redirect("/Users");
-      res.json("success")
+      res.json(dbUser)
     }).catch(function(err){
+     
       res.status(500).json(err);
     });
     
@@ -76,7 +74,7 @@ router.get("/users/:id", function(req, res) {
   }).then(function(dbUser) {
     console.log(dbUser);
     // res.json(dbUser, `retrieved user with id: ${req.params.id}`);
-    res.json("testing response success of user findOne")
+    res.json( dbUser)
     //TODO: check in with nicole where this will route to, pulling functional record correctly based on user id.
   }).catch(function(err){
     res.status(500).json(err);
@@ -94,7 +92,7 @@ router.put("/users/update/:id", function(req, res) {
   ).then(function(dbUsers) {
     //this should redirect us to homepage, may have to edit.
     // res.json("/");
-    res.json("test for user update, nice job");
+    res.json(dbUsers);
   }).catch(function(err){
     res.status(500).json(err);
   });
@@ -112,105 +110,7 @@ router.delete("/users/delete/:id", function (req, res) {
     res.status(500).json(err);
   });
 });
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~HERE END THE ROUTES FOR USERS~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-
-
-//==========================HERE BEGIN THE TODO: customer ROUTES======================================
-
-router.post("/customers/create", function (req, res) {
-  db.Customer.create({
-    UserId : req.body.id
-  }).then(function (dbCustomer) {
-        console.log(dbCustomer);
-        // res.redirect("/customer");
-        //TODO: link this properly with the frontend, both on req.body having an id key (or whatever it's called)
-        res.json("we made you that customer");
-      }).catch(function(err){
-        res.status(500).json(err);
-      });
-    });
-
-router.get("/customers", function (req,res) {
-  db.Customer.findAll({
-    include: [db.User]
-  }).then(function (dbCustomer){
-    console.log(dbCustomer)
-    //   //TODO: make sure this render is correct
-    //   return res.render("index", allCustomer);
-    return res.json(dbCustomer)
-  }).catch(function(err){
-    res.status(500).json(err);
-  });
-})
-
-
-router.get("/customers/:id", function (req,res) {
-  db.Customer.findOne({
-    where: {id: req.params.id}
-  })
-  .then(function(dbCustomer) {
-    const dbCustomerJson = (dbCustomer=>Customer.toJSON());
-    var oneCustomer = { Customer: dbCustomerJson };
-    //TODO: make sure this render is correct
-    return res.render("index", oneCustomer);
-  })
-  .catch(function(err){
-    res.status(500).json(err);
-  });
-})
-
-
-
-
-//==========================HERE END THE customer ROUTES======================================
-
-
-//================================HERE BEGIN THE PROVIDER ROUTES======================================
-router.post("/providers/create", function (req, res) {
-    db.Provider.create({
-      UserId : req.body.id
-    }).then(function (dbProvider) {
-          console.log(dbProvider);
-          // res.redirect("/provider");
-          //TODO: link this properly with the frontend, both on req.body having an id key (or whatever it's called)
-          res.json("we made you that provider");
-        }).catch(function(err){
-          res.status(500).json(err);
-        });
-      });
-
-  router.get("/providers", function(req, res) {
-
-  db.Provider.findAll({})
-    .then(function(dbProvider) {
-      console.log({dbProvider});
-      //TODO: make sure this returns what frontend needs
-      return res.json(dbProvider);
-    }).catch(function(err){
-      console.log("THIS IS THE ERROR ON 191", err)
-      res.status(500).json(err);
-    });
-});
-
-router.get("/providers/:id", function (req,res) {
-  db.Provider.findOne({
-    where: {id: req.params.id}
-  })
-  .then(function(dbProvider) {
-    console.log(dbProvider)
-    
-   
-    //TODO: make sure we are rendering properly or sending back to the right place for frontend
-    return res.json({provider: dbProvider})
-    // return res.render("index", {provider: dbProvider});
-  })
-  .catch(function(err){
-    res.status(500).json(err);
-  });
-})
-
-//==============================HERE END THE PROVIDER ROUTES===========================================
 
 
 
@@ -218,7 +118,7 @@ router.get("/providers/:id", function (req,res) {
 //offer_posts CREATE NEW POST 
 router.post("/offer_posts/create", function(req,res) {
   console.log(req.body)
-  db.offer_post.create(req.body)
+  db.Post.create(req.body)
   .then(function(dbPost) {
       console.log(dbPost);
       //TODO: decide where this redirect will go, ask Nicole. for now will refresh the total posts page.
@@ -232,7 +132,7 @@ router.post("/offer_posts/create", function(req,res) {
 // //offer_posts READ ALL 
 router.get("/offer_posts", function(req,res) {
   
-  db.offer_post.findAll()
+  db.Post.findAll()
   
   .then(function(dbPost) {
     console.log(dbPost);
@@ -245,50 +145,11 @@ router.get("/offer_posts", function(req,res) {
   });
 });
 
-// router.get("/offer_posts/:range", function(req,res) {
-//   let rangeRel = req.params.range;
-//   if (0 < rangeRel < 2) {
-//     db.offerPost
-//       .findAll({
-//         where: {
-//           //TODO: the syntax for this is really funky -
-//           // might need the op. and followed by a nest op, looking at the between operator for lat and long between the current value +-.1
-//         },
-//       })
-//       .then(function (dbPost) {
-//         console.log(dbPosts);
-//         const dbPostsJson = dbPosts.map((post) => post.toJSON());
-//         let hbrsObj = { offer_post: dbPostsJson };
-//         // return res.render("offer_post", hbrsObj);
-//         //TODO: convert the below to above line when post.handlebars exists
-//         return res.json(hbrsObj);
-//       });
-//   }
-//   if (2 < rangeRel < 5) {
-//     db.offerPost
-//     //TODO: work on the joins involved this has joins written all over it 
-//       .findAll({
-//         where: {
-//           //the syntax for this is really funky -
-//           // might need the op. and followed by a nest op, looking at the between operator for lat and long between the current value +-.1 this ultimately depends on how lat and lon are taken in, displayed, and where.
-//         },
-//       })
-//     .then(function (dbPost) {
-//       console.log(dbPosts);
-//       const dbPostsJson = dbPosts.map((post) => post.toJSON());
-//       let hbrsObj = { offer_post: dbPostsJson };
-//       // return res.render("offer_post", hbrsObj);
-//       //TODO: convert the below to above line when post.handlebars exists
-//       return res.json(hbrsObj);
-//     });
-//   }
-// });
-
 
 // // offer_posts UPDATE, by post id. 
 router.put("/offer_posts/update/:id", function (req,res) {
 
-  db.offer_post.update({
+  db.Post.update({
     title: req.body.title,
     text: req.body.text,
     size_restrictions: req.body.size_restrictions,
@@ -314,7 +175,7 @@ router.put("/offer_posts/update/:id", function (req,res) {
 
 // offer_posts DELETE, by post id.
 router.delete("/offer_posts/:id", function (req, res) {
-  db.offer_post
+  db.Post
     .destroy({
       where: {
         id: req.params.id,
@@ -488,7 +349,7 @@ router.put("/ratings/update/:id", function (req,res) {
 });
 
 router.delete("/ratings/delete/:id", function (req, res) {
-  db.User.destroy({
+  db.Rating.destroy({
     where: {
       id: req.params.id
     }
