+++ models/offer_posts.js
@@ -1,18 +1,13 @@
 module.exports = function(sequelize, DataTypes) {
-    var Post = sequelize.define("offer_post", {
+    var Post = sequelize.define("Post", {
       title: {
           type:DataTypes.STRING,
           allowNull: false,
-          validate: {
-            len: [1]
-          }
       },
       text: {
           type:DataTypes.TEXT,
           allowNull: false,
-          validate: {
-            len: [1, 1000]
-          }
+
       },
       animal_type: {
           type:DataTypes.STRING,
@@ -25,42 +20,30 @@ module.exports = function(sequelize, DataTypes) {
       duration: {
         type:DataTypes.INTEGER,
         allowNull: false,
-        validate: {
-            len: [1]
-        }
+
       },
       range: {
         type:DataTypes.INTEGER,
         allowNull: false,
-        validate: {
-            len: [1]
-        }
+ 
       },
       picture: {
         type:DataTypes.STRING,
         allowNull: true,
-        validate: {
-            isUrl: true
-        }
+
       },
       service_type: {
         type:DataTypes.STRING,
         allowNull: false,
-        validate: {
-            len: [1]
-        }
+
       },
 
     });  
     Post.associate = function(models) {
-        Post.belongsTo(models.Provider, {
-          foreignKey: {
-            allowNull: false
-          }
+        Post.belongsTo(models.User);
+        Post.belongsTo(models.User, {
+          as: "Provider"
         });
-        Post.hasMany(models.Rating, {
-          onDelete: "cascade"
-        })
       };
     return Post;
   };
