+++ models/pets.js
@@ -1,27 +1,12 @@
-//const { Sequelize } = require(".");
-
 module.exports = function(sequelize, DataTypes) {
     var Pet = sequelize.define("Pet", {
       first_name: {
           type:DataTypes.STRING,
           allowNull: false,
-          validate: {
-            len: [1]
-          }
-      },
-      last_name: {
-          type:DataTypes.STRING,
-          allowNull: true,
-          validate: {
-            len: [1]
-          }
       },
       special_care: {
         type:DataTypes.STRING,
         allowNull: true,
-        validate: {
-            len: [1]
-        }
     },
       pet_type: {
           type:DataTypes.STRING,
@@ -30,9 +15,6 @@ module.exports = function(sequelize, DataTypes) {
        breed: {
           type:DataTypes.STRING,
           allowNull: false,
-          validate: {
-            len: [1],
-          }
         },
       size: {
         type:DataTypes.STRING,
@@ -41,35 +23,22 @@ module.exports = function(sequelize, DataTypes) {
       temperment: {
         type:DataTypes.STRING,
         allowNull: false,
-        validate: {
-            len: [1]
-        }
       },
       picture: {
         type:DataTypes.STRING,
         allowNull: true,
-        validate: {
-            isUrl: true
-        }
       },
       age: {
         type:DataTypes.INTEGER,
         allowNull: false,
-        validate: {
-            max: 23, 
-            min: 0
-        }
+
       },
     });  
     Pet.associate = function(models) {
-        Pet.belongsTo(models.Customer, {
-          foreignKey: {
-            allowNull: false
-          }
+        Pet.belongsTo(models.User);
+        Pet.belongsTo(models.User, {
+          as: "Customer"
         });
-        Pet.hasMany(models.Rating, {
-          onDelete: "cascade"
-      })
       };
     return Pet;
   };
